refactor(D-170 C): extract Pydantic contracts from probability_skill into probability_models
1,685→1,462 lines. The 11 contract classes (ProbabilityInput → ProbabilityFailure)
+ the DOCTRINE_VERSION stamp now live in probability_models.py (contract vs engine
separation, SRP). probability_skill re-exports everything (noqa: F401) so every
existing importer — UI streaming contracts, tests, the brain — is untouched
(D-166/D-168 backward re-export pattern). Verbatim move, zero behavior change:
live engine smoke C(11,3)=165 / same=14 / doctrine 1.5.0 via the re-export.
